Espressif IDF v5 support (#1)
* Fix null terminated strings in Parameters

* Use network instead of WLAN

* Change start order and use connected wifi to start homespan

* IDF v5 support

+++ src/Dimmer/KnxChannelDimmer.cpp
@@ -70,8 +70,8 @@ void KnxChannelDimmer::commandPower(DimmerBridge* dimmerBridge, bool power)
 
 void KnxChannelDimmer::setup()
 {
-    koSetWithoutSend(KO_DIMMER, 0);
-    koSetWithoutSend(KO_DIMMER_FEEDBACK, 0);
+    koSetWithoutSend(KO_DIMMER, (uint8_t) 0);
+    koSetWithoutSend(KO_DIMMER_FEEDBACK,(uint8_t) 0);
     koSendReadRequest(KO_DIMMER_FEEDBACK);
 }
